refactor: auth 분리, dtos.d.ts->types.d.ts 변경
