+++ web/src/app/api/lib/clientAuth.ts
@@ -1,0 +1,6 @@
+import { getAuth as clientGetAuth } from '@firebase/auth';
+import clientApp from '@/app/api/config/firebase.client';
+
+const clientAuth = clientGetAuth(clientApp);
+
+export default clientAuth;
