+++ web/src/app/api/repository/users/createUser.ts
@@ -1,11 +1,11 @@
-import { RegisterDto } from '@/app/api/lib/dtos';
+import { RegisterDto } from '@/app/api/lib/types';
 import {
   EMAIL_ALREADY_EXISTS,
   INVALID_DISPLAY_NAME,
   INVALID_EMAIL,
   WEAK_PASSWORD,
 } from '@/app/api/const/errors';
-import { auth } from '@/app/api/lib/getAuth';
+import auth from '@/app/api/lib/auth';
 
 const createUser = async ({ email, password, nickName }: RegisterDto) => {
   const user = await auth
@@ -19,6 +19,7 @@ const createUser = async ({ email, password, nickName }: RegisterDto) => {
         throw new Error('이미 존재하는 이메일입니다');
       if (err.code === INVALID_EMAIL)
         throw new Error('유효하지 않은 이메일입니다');
+
       if (err.code === WEAK_PASSWORD)
         throw new Error('비밀번호 형식이 올바르지 않습니다.');
       if (err.code === INVALID_DISPLAY_NAME)
